+++ src/app/(routes)/solved/components/center/common-interview-questions/index.tsx
